Timer now works. WPM still not implemented.
This commit adds a working timer. Words per minute counter is still not
yet implemented.

Next commit should focus on getting WPM to work, and changing the
gameplay mechanics so that the player is forced to backspace in order to
fix typos, instead of the cursor not moving when the keypress is
incorrect. This change would more-so reflect actual typing than it does
in the current state.